+++ src/alignment/phmm/traits.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "fuzzing")]
 pub mod float_compare {
     use crate::{
-        alignment::phmm::{EmissionParams, GlobalPhmm, LayerParams, TransitionParams},
+        alignment::phmm::{CorePhmm, EmissionParams, GlobalPhmm, LayerParams, TransitionParams},
         math::{NearlyEqual, NearlyEqualMethod},
     };
 
@@ -32,10 +32,16 @@ pub mod float_compare {
         }
     }
 
+    impl<T: NearlyEqual<T> + Copy, const S: usize> NearlyEqual<T> for CorePhmm<T, S> {
+        fn nearly_equal<M: NearlyEqualMethod<T>>(&self, b: &Self, strategy: &M) -> (bool, Option<(T, T)>) {
+            self.0.nearly_equal(&b.0, strategy)
+        }
+    }
+
     impl<T: NearlyEqual<T> + Copy, const S: usize> NearlyEqual<T> for GlobalPhmm<T, S> {
         fn nearly_equal<M: NearlyEqualMethod<T>>(&self, b: &Self, strategy: &M) -> (bool, Option<(T, T)>) {
             if self.mapping == b.mapping {
-                self.params.nearly_equal(&b.params, strategy)
+                self.core.nearly_equal(&b.core, strategy)
             } else {
                 (false, None)
             }